+++ include/swift/Frontend/ParseableInterfaceSupport.h
@@ -42,7 +42,7 @@ extern version::Version InterfaceFormatVersion;
 llvm::Regex getSwiftInterfaceFormatVersionRegex();
 llvm::Regex getSwiftInterfaceModuleFlagsRegex();
 
-/// Emit a stable, parseable interface for \p M, which can be used by a client
+/// Emit a stable module interface for \p M, which can be used by a client
 /// source file to import this module, subject to options given by \p Opts.
 ///
 /// Unlike a serialized module, the textual format generated by
