+++ lib/Frontend/CMakeLists.txt
@@ -7,6 +7,7 @@ add_swift_host_library(swiftFrontend STATIC
   Frontend.cpp
   FrontendInputsAndOutputs.cpp
   FrontendOptions.cpp
+  ModuleInterfaceBuilder.cpp
   ParseableInterfaceModuleLoader.cpp
   ParseableInterfaceSupport.cpp
   PrintingDiagnosticConsumer.cpp
